authorization and authentication

+++ app/Http/Controllers/PostController.php
@@ -39,6 +39,7 @@ class PostController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Post::class);
         $categories = Category::all();
         return view('create', compact('categories'));
     }
@@ -48,6 +49,8 @@ class PostController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Post::class);
+
         $request->validate([
             'image' => ['required', 'max: 2028', 'image'],
             'title' => ['required', 'max: 255'],
@@ -86,6 +89,8 @@ class PostController extends Controller
     public function edit(string $id)
     {
         $post = Post::findorFail($id);
+        $this->authorize('update', $post);
+
         $categories = Category::all();
         return view('edit', compact('post', 'categories'));
     }
@@ -95,6 +100,9 @@ class PostController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $post = Post::findorFail($id);
+
+        $this->authorize('update', $post);
         $request->validate([
             'title' => ['required', 'max: 255'],
             'category_id' => ['required', 'integer'],
@@ -102,7 +110,7 @@ class PostController extends Controller
         ]);
 
 
-        $post = Post::findorFail($id);
+
 
 
         if($request->hasFile('image')){
@@ -137,12 +145,15 @@ class PostController extends Controller
     public function destroy(string $id)
     {
         $post = Post::findorFail($id);
+        $this->authorize('delete', $post);
+
         $post->delete();
         return redirect()->route('posts.index');
     }
 
     public function trashed()
     {
+        //$this->authorize('delete', Post::class);
         $posts = Post::onlyTrashed()->get();
         return view('trashed', compact('posts'));
     }
@@ -150,6 +161,9 @@ class PostController extends Controller
     public function restore($id)
     {
         $post = Post::onlyTrashed()->findOrFail($id);
+        $this->authorize('delete', $post);
+
+
         $post->restore();
 
         return redirect()->back();
@@ -158,6 +172,9 @@ class PostController extends Controller
     public function forceDelete($id)
     {
         $post = Post::onlyTrashed()->findOrFail($id);
+        $this->authorize('delete', $post);
+
+
 
         File::delete(public_path($post->image));
